Commit of initial project state
This test project was initially a quick day long endeavor for experimentation, but is not in a repo but further development. Current state includes a BST implementation derived from a more generic Tree class. Each node has a key and a generic value in the form of a tuple. Visuals were implemented using turtle.py but will promptly be swapped out for tkinter shape visual functionality as each shape can be given an id to be later manipulated.