Changed graphics implementation
In place of using the turtle.py module, the tkinter's native functions for drawing shapes have been used since they have be referenced and changed later through the use of ids. All is stable but nodes current overlap if they are the right node of a left parent and the left node of a right parent.